requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyMuPDF 
python-multipart
python-dotenv